requirements.txt
alembic==1.15.1
asttokens==3.0.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
findspark==2.0.1
fsspec==2025.2.0
greenlet==3.1.1
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.7
Pygments==2.19.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
six==1.17.0
spark-nlp==5.5.3
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
apify-client==1.9.2
asyncio==3.4.3
aiohttp==3.11.14
matplotlib==3.10.1
seaborn==0.13.2
streamlit==1.43.2
scipy==1.15.2
pyarrow==19.0.1
watchdog==6.0.0
pybind11>=2.12.0 
getenv==0.2.0
python-dotenv==1.0.1